+++ parser/sql/statement/type/mysql/src/main/java/org/apache/shardingsphere/sql/parser/statement/mysql/ddl/MySQLAlterViewStatement.java
@@ -35,7 +35,7 @@ public final class MySQLAlterViewStatement extends AlterViewStatement implements
     private String viewDefinition;
     
     @Override
-    public Optional<SelectStatement> getSelectStatement() {
+    public Optional<SelectStatement> getSelect() {
         return Optional.ofNullable(select);
     }